pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-coder-llm"
version = "0.1.0"
description = "Medical coding CLI pipeline (Python port of with-bun)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "python-dotenv>=1.0",
]

[project.scripts]
medical-coder-llm = "medical_coder_llm.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
